+++ WebApplication/src/orthancApi.js
@@ -739,16 +739,15 @@ export default {
         // stl/app/o3dv.html
         // stl/app/three.html
     },
-   getWeasisViewerUrl(level, resourceDicomUid) {
-        const parts = ["$dicom:rs", "--url", `"${window.location.origin}/dicom-web"`, "-r", `"studyUID=${resourceDicomUid}"`];
+   getWeasisViewerUrl(resourceDicomUid) {
+        const dicomWebRootUrl = new URL(orthancApiUrl + "dicom-web", window.location.origin);
+        const parts = ["$dicom:rs", "--url", `"${dicomWebRootUrl.toString()}"`, "-r", `"studyUID=${resourceDicomUid}"`];
         const url = "weasis://?" + parts.map(v => encodeURIComponent(v)).join("+");
         return url
     },
     getWeasisViewerUrlForBulkStudies(studiesDicomIds) {
         const studyUids = studiesDicomIds.join(",");
-        const parts = ["$dicom:rs", "--url", `"${window.location.origin}/dicom-web"`, "-r", `"studyUID=${studyUids}"`];
-        const url = "weasis://?" + parts.map(v => encodeURIComponent(v)).join("+");
-        return url
+        return this.getWeasisViewerUrl(studyUids);
     },
     getInstancePreviewUrl(orthancId) {
         return orthancApiUrl + "instances/" + orthancId + "/preview";